tools/models-codegen/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.5.0
anyio==3.7.1
argcomplete==3.1.1
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
datamodel-code-generator==0.21.3
dnspython==2.4.1
email-validator==2.0.0.post2
genson==1.2.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflect==5.6.2
isort==5.12.0
jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.6
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
openapi-schema-validator==0.4.4
openapi-spec-validator==0.5.2
packaging==23.1
pathable==0.4.3
pathspec==0.11.2
platformdirs==3.10.0
prance==23.6.21.0
pydantic==2.1.1
pydantic-core==2.4.0
pyrsistent==0.19.3
PySnooper==1.2.0
python-benedict==0.32.0
python-fsutil==0.10.0
python-slugify==8.0.1
pyyaml==6.0.1
requests==2.31.0
rfc3339-validator==0.1.4
ruamel-yaml==0.17.32
ruamel-yaml-clib==0.2.7
six==1.16.0
sniffio==1.3.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.7.1
urllib3==2.0.4
